+++ types/GeoJsonDumper.d.ts
@@ -3,7 +3,7 @@ import AdminLevel, { AdminLevelObject } from "./AdminLevel";
 export interface GeoJson {
     readonly type: "Feature";
     readonly properties: {
-        readonly [property: string]: string | string[] | number | AdminLevel[] | AdminLevelObject[] | undefined;
+        readonly [property: string]: string | string[] | number | boolean | AdminLevel[] | AdminLevelObject[] | undefined;
     };
     readonly geometry: {
         readonly type: "Point";